move_to_check_config
